Changed some hard to read verbage for availabilty replicas
Changed "A given instance can host only one availability replica per availability group" to "A given instance can only act as one availability replica per availability group" as an instance *is* a replica, having more than one availability group does not make it a new or multiple replica. Using "hosted" made it sound as though it was something that could be changed or transferred when it cannot. It either is or is not a replica and cannot be migrated or moved or "hosted" on any other replica. Hence the change to "can only act as one".